requirements.txt
flask==2.0.1
google-generativeai==0.3.2
python-json-logger==2.0.7
jsonrpcserver==5.0.9
python-dotenv==0.19.0
werkzeug==2.0.1
jsonschema==3.2.0
pyjwt==2.1.0 